+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
+
+**v0.1.2 - 09-07-2026**
+
+- Passphrases and ChalResp hardening
+- Improved import identity
+- Fixed id removal & hw key delete gate
+
+
 **v0.1.1 - 09-05-2026**
 
 - Cross-platform keychain fix